tests/test_case_upstream_lambda/pipeline_code/requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-aws-lambda
opentelemetry-instrumentation-botocore
opentelemetry-instrumentation-requests
requests
